Refactor/#1432 onboarding
