+++ docs/nextjs_architecture_guidelines.md
@@ -41,6 +41,7 @@ src/
 ルール:
 - ✅ インタラクティブな要素（クリック、フォーム、useState/useEffect など）を持つ場合は `'use client'` を原則として付与する
 - ✅ 静的な表示のみの場合は `import 'server-only'` を明示してServer Component を推奨する。
+- ✅ **Props型,State型定義はコンポーネントファイル内に記載する**（外部ファイルでの定義禁止）
 - ❌ 直接データベースアクセス禁止
 - ℹ️ 備考：
   サーバーコンポーネントはデフォルトでサーバー実行されますが、明示的にサーバー処理であることを示すために必要に応じて `import 'server-only'` を使用する。
@@ -97,7 +98,11 @@ src/
 実行環境: 共通（環境非依存）
 
 ルール:
-- ✅ 環境非依存の型定義のみ
+- ✅ 環境非依存の型定義（サーバー・クライアント両方で使用可能）
+- ✅ ビジネスロジック・ドメインモデルに関する型定義
+- ✅ API レスポンス・リクエストの型定義
+- ✅ 複数のコンポーネント・関数で共有される型定義
+- ❌ **Props型,State型定義の記載禁止**（コンポーネントファイル内に記載すること）
 - ❌ 実行時コード・関数の実装禁止
 
 ### 🛠️ src/features/[feature-name]/utils/
